Refactoring. Spostati i Models nel progetto frontend. Eliminato il progetto Shared che li conteneva precedentemente.
