--- src/Gdn.Presentation.Shared/Models/ErrorViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Gdn.Presentation.Shared.Models;
-
-public class ErrorViewModel
-{
-    public string Code { get; set; } = string.Empty;
-    public string Description { get; set; } = string.Empty;
-}
