+++ src/Gdn.Web.Fluentblazor/Components/TaxRates/EditTaxRateDialog.razor
@@ -74,7 +74,6 @@
     IEnumerable<TaxRateNatureViewModel>? _taxRateNatures;
     List<Option<int>>? _taxRateNatureOptions;
     Option<int>? _selectedTaxRateNatureOption;
-    string? _taxRateNatureValue;
 
     protected override async Task OnInitializedAsync()
     {
