+++ src/Gdn.Web.Fluentblazor/Components/_Imports.razor
@@ -17,10 +17,10 @@
 @using Gdn.Web.Fluentblazor.Components.Invoices
 @using Gdn.Web.Fluentblazor.Components.Custom
 @using Gdn.Web.Fluentblazor.Extensions
-@using Gdn.Web.Fluentblazor.Models
 @using Gdn.Web.Fluentblazor.Utils
 
-@using Gdn.Presentation.Shared.Models.TaxRates
-@using Gdn.Presentation.Shared.Models.Customers
-@using Gdn.Presentation.Shared.Models.MeasurementUnits
-@using Gdn.Presentation.Shared.Models.Invoices
+@using Gdn.Web.Fluentblazor.Models
+@using Gdn.Web.Fluentblazor.Models.TaxRates
+@using Gdn.Web.Fluentblazor.Models.Customers
+@using Gdn.Web.Fluentblazor.Models.MeasurementUnits
+@using Gdn.Web.Fluentblazor.Models.Invoices
