+++ src/Gdn.Web.Fluentblazor/Models/TaxRates/TaxRateNatureViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Gdn.Presentation.Shared.Models.TaxRates;
+﻿namespace Gdn.Web.Fluentblazor.Models.TaxRates;
 
 public class TaxRateNatureViewModel
 {
