+++ src/Gestionale DotNet.sln
@@ -15,8 +15,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gdn.Application", "Gdn.Appl
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gdn.Persistence", "Gdn.Persistence\Gdn.Persistence.csproj", "{E8B1D26B-F909-4815-B36A-0615FFD9B10A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gdn.Presentation.Shared", "Gdn.Presentation.Shared\Gdn.Presentation.Shared.csproj", "{686EA417-A961-4E03-BB46-39DB011E4022}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Frontend", "Frontend", "{ADB80CAD-2484-4BDA-8388-1992EC086A60}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gdn.Web.Fluentblazor", "Gdn.Web.Fluentblazor\Gdn.Web.Fluentblazor.csproj", "{6E8650DE-DDEE-4009-B079-41EA77DBAE89}"
@@ -49,10 +47,6 @@ Global
 		{E8B1D26B-F909-4815-B36A-0615FFD9B10A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8B1D26B-F909-4815-B36A-0615FFD9B10A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E8B1D26B-F909-4815-B36A-0615FFD9B10A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{686EA417-A961-4E03-BB46-39DB011E4022}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{686EA417-A961-4E03-BB46-39DB011E4022}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{686EA417-A961-4E03-BB46-39DB011E4022}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{686EA417-A961-4E03-BB46-39DB011E4022}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6E8650DE-DDEE-4009-B079-41EA77DBAE89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6E8650DE-DDEE-4009-B079-41EA77DBAE89}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E8650DE-DDEE-4009-B079-41EA77DBAE89}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -81,7 +75,6 @@ Global
 		{7D454E9E-EE85-4419-BFA3-44151273C6B7} = {4B6E43B0-68E0-4D3B-A40A-D74E8373106C}
 		{8463897E-F77F-4D33-BD9A-E6EFC7DF0F5C} = {4B6E43B0-68E0-4D3B-A40A-D74E8373106C}
 		{E8B1D26B-F909-4815-B36A-0615FFD9B10A} = {79530925-09FE-42D3-9B8D-403AEE1978BA}
-		{686EA417-A961-4E03-BB46-39DB011E4022} = {E5CE911E-49D2-4590-9F45-8B800740722C}
 		{ADB80CAD-2484-4BDA-8388-1992EC086A60} = {E5CE911E-49D2-4590-9F45-8B800740722C}
 		{6E8650DE-DDEE-4009-B079-41EA77DBAE89} = {ADB80CAD-2484-4BDA-8388-1992EC086A60}
 		{0972B11B-511A-43AB-8F77-914AEB238E51} = {0020C031-90C3-41E5-B517-D2D6CED40D68}
